Embedded messages handling (partial)
Embedded messages aren't part of the .pst btree structure so some
properties are missing (filename, creation time etc.).

The first NullPointerException was at line 222 (attach.getCreationTime()
returns 0).

Embedded messages are now parsed. The extracted artifacts are added to
the results and then to the blackboard.

Please note that the embedded message itself is not saved to disk at the
moment (a placeholder is saved) but its attachments, if any, are parsed
too.