5333: Allow training data that contains only one category.
The way this works is that every time we train, we include an empty document for each category.

I noticed that classification now depends on order.
If you load the Nonnotable documents of the motorcycle test first, you get a different model
then if you load them in a random order or with notable documents first.
I'm still looking into that.