Update VMExtractorIngestModule.java
Moved defining fileTypeDetector outside of loop so it only happens once.
If there is an error initializing fileTypeDetector then add the file to the list of VM's that we look at even if it may not be a VM, this way we do not loose it.
Continue on in the loop skipping over a potential NPE in the code after the initilization.